rename cv-card directive to mb-dashcard and move the code from the card.directives into dashboard.directives.  this makes it more clear that this code is only used for rendering dashboard cards now.
