+++ resources/frontend_client/app/dashboard/partials/dash_view.html
@@ -62,7 +62,7 @@
                             <a class="DashCard-remove" ng-click="removeCard(index)">
                                 <mb-icon name="close" class="DashCard-removeIcon"  width="12px" height="12px"></mb-icon>
                             </a>
-                            <cv-card card="dashcard.card" card-data="dashcard.cardData.data" card-settings="cardSettings" index="index" wait-for-gridster />
+                            <mb-dashcard card="dashcard.card" card-data="dashcard.cardData.data" card-settings="cardSettings" index="index" wait-for-gridster />
                         </li>
                     </ul>
                 </div>
